requirements.txt
langchain
pypdf
unstructured
sentence_transformers
pinecone-client
#llama-cpp-python
huggingface_hub
IProgress 
accelerate